# ochap-authentication/AuthServerSwagger.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"OpenAPI definition",
      "version":"v0"
   },
   "servers":[
      {
         "url":"http://localhost:8081",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/auth/login":{
         "post":{
            "tags":[
               "Auth"
            ],
            "summary":"Log in",
            "operationId":"login",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AuthLoginCommand"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successfully logged in",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/TokenDTO"
                        }
                     }
                  }
               },
               "403":{
                  "description":"The credentials are incorrect",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/Error"
                        }
                     }
                  }
               }
            }
         }
      },
      "/accounts/register":{
         "post":{
            "tags":[
               "Account"
            ],
            "summary":"Register an Account",
            "operationId":"create",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AccountRegisterCommand"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "201":{
                  "description":"Account created",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/AccountInfoDTO"
                        }
                     }
                  }
               },
               "422":{
                  "description":"The data you entered are not valid",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorsDTO"
                        }
                     }
                  }
               },
               "409":{
                  "description":"The username already exist",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/Error"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "AccountInfoDTO":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"integer",
                  "format":"int64"
               },
               "username":{
                  "type":"string"
               },
               "role":{
                  "type":"string"
               }
            }
         },
         "TokenDTO":{
            "type":"object",
            "properties":{
               "token":{
                  "type":"string"
               },
               "account":{
                  "$ref":"#/components/schemas/AccountInfoDTO"
               }
            }
         },
         "Error":{
            "type":"object",
            "properties":{
               "error":{
                  "type":"string"
               }
            }
         },
         "AuthLoginCommand":{
            "type":"object",
            "properties":{
               "username":{
                  "type":"string"
               },
               "password":{
                  "type":"string"
               }
            }
         },
         "ErrorDTO":{
            "type":"object",
            "properties":{
               "property":{
                  "type":"string"
               },
               "message":{
                  "type":"string"
               }
            }
         },
         "ErrorsDTO":{
            "type":"object",
            "properties":{
               "errors":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/ErrorDTO"
                  }
               }
            }
         },
         "AccountRegisterCommand":{
            "type":"object",
            "properties":{
               "username":{
                  "type":"string"
               },
               "password":{
                  "type":"string"
               }
            }
         }
      },
      "securitySchemes":{
         "bearer":{
            "type":"http",
            "scheme":"bearer",
            "bearerFormat":"JWT"
         }
      }
   }
}